+++ frontend/src/redux/store.js
@@ -11,12 +11,12 @@ const reducers = combineReducers({
     inbox: inboxReducer, 
 });
 
-  const persistConfig = {
+const persistConfig = {
     key: 'root',
     storage,
 };
 
-  const persistedReducer = persistReducer(persistConfig, reducers);
+const persistedReducer = persistReducer(persistConfig, reducers);
 
 const store = configureStore({
     reducer: persistedReducer,
